requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
Django==3.2.5
django-cors-headers==3.13.0
django-embed-video==1.4.4
django-environ==0.8.1
django-zxcvbn-password==2.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.3
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.9
python-http-client==3.3.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
zxcvbn==4.4.28
